game_environment.py
```python
class GameEnvironment:
    def __init__(self):
        self.environment = None

    def generate(self, ai):
        """
        Use AI to generate a dynamic game environment.
        """
        self.environment = ai.generate()

    def start(self, character):
        """
        Start the game with the given character.
        """
        if self.environment is None:
            raise Exception("Game environment not generated. Use generate() method to generate the environment.")

        # Add character to the environment
        self.environment.add_character(character)

        # Start the game loop
        while True:
            # Update the game environment
            self.environment.update()

            # Check if the character is still alive
            if not character.is_alive():
                print("Game Over")
                break

            # Check if the character has reached the goal
            if character.has_reached_goal():
                print("Congratulations! You have reached the goal.")
                break
```
